+++ examples/parametric_models/04-parametric-models.py
@@ -24,10 +24,10 @@ def example_parametric_models_04():
                 if len(transition.value().gather_variables()) > 0:
                     new_var = pycarl.Variable("p{}".format(i))
                     i += 1
-                    new_pol = stormpy.Polynomial(new_var)
-                    pol_in_right_format = stormpy.FactorizedPolynomial(new_pol, transition.value().numerator.cache())
+                    new_pol = stormpy.RawPolynomial(new_var)
+                    pol_in_right_format = stormpy.Polynomial(new_pol, transition.value().numerator.cache())
 
-                    new_factorized_ratfunc = stormpy.FactorizedRationalFunction(pol_in_right_format)
+                    new_factorized_ratfunc = stormpy.RationalFunction(pol_in_right_format)
                     transition.set_value(new_factorized_ratfunc)
 
     # Display