+++ tests/core/test_core.py
@@ -7,7 +7,7 @@ class TestCore:
 
         rational = stormpy.Rational(0.25)
         assert str(rational) == "1/4"
-        pol1 = stormpy.FactorizedPolynomial(32)
-        pol2 = stormpy.FactorizedPolynomial(2)
-        rat = stormpy.FactorizedRationalFunction(pol1, pol2)
+        pol1 = stormpy.Polynomial(32)
+        pol2 = stormpy.Polynomial(2)
+        rat = stormpy.RationalFunction(pol1, pol2)
         assert str(rat) == "16"